Test for the methods computeTax, weeklySalary, fortnightSalary and computeKiwiSaver successful.

+++ Assignment3/IT7320_Assignment3_2183490/src/main/java/Employee.java
@@ -6,6 +6,7 @@
  */
 public class Employee {
 
+		
 		private int salary;
 		
 		private double kiwisaverPercentage;
@@ -13,7 +14,7 @@ public class Employee {
 		/**
 		 * Default constructor.
 		 */
-		public Employee(int salary, int kiwisaverPercentage) {
+		public Employee(int salary, double kiwisaverPercentage) {
 			this.salary = salary;
 			this.kiwisaverPercentage = kiwisaverPercentage;
 		}
